backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
scikit-learn>=1.4.0
joblib>=1.3.2
psycopg2-binary>=2.9.7
numpy>=1.24.3
pandas>=2.0.0
pydantic[email]>=2.4.2
bcrypt>=4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
